Fix build error due to type conversion in std::min

+++ extras/rapidfuzz_amalgamated.hpp
@@ -3491,7 +3491,7 @@ static inline FlaggedCharsWord flag_similar_characters_word(
     uint64_t BoundMask = (1ull << 1 << Bound) - 1;
 
     int j = 0;
-    for (; j < std::min(Bound, T.size()); ++j)
+    for (; j < std::min(Bound, (uint64_t)T.size()); ++j)
     {
         uint64_t PM_j = PM.get(T[j]) & BoundMask & (~P_flag);
